chore: rename components on auth folders
