+++ app/(root)/page.tsx
@@ -1,12 +1,12 @@
-import { HeroComponent } from '@/components/root/hero-component'
-import { AboutComponent } from '@/components/root/about-component'
+import { Hero } from '@/components/root/hero'
+import { About } from '@/components/root/about'
 
 const Page = () => {
   return (
     <div className='flex flex-col min-h-[87dvh]'>
       <main className='flex-1'>
-        <HeroComponent />
-        <AboutComponent />
+        <Hero />
+        <About />
       </main>
     </div>
   )
