+++ components/root/about.tsx
@@ -1,6 +1,6 @@
 import { HeadphonesIcon, TicketIcon, InfoIcon } from 'lucide-react'
 
-const AboutComponent = () => {
+const About = () => {
   return (
     <section className='w-full py-12 md:py-24 lg:py-32'>
       <div className='container space-y-12 px-4 md:px-6'>
@@ -35,4 +35,4 @@ const AboutComponent = () => {
   )
 }
 
-export { AboutComponent }
+export { About }
