+++ components/root/auth/signin-form.tsx
@@ -1,33 +1,30 @@
 'use client'
 
 import { useState } from 'react'
+import Link from 'next/link'
 import { useForm, SubmitHandler } from 'react-hook-form'
 import { EyeIcon, EyeOffIcon, AlertCircle } from 'lucide-react'
 
 import { Label } from '@/components/ui/label'
 import { Input } from '@/components/ui/input'
 import { Button } from '@/components/ui/button'
+import { Checkbox } from '@/components/ui/checkbox'
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert'
+import { useSignin } from '@/hooks/auth/use-signin'
+import { AuthSignInFormType } from '@/lib/definitions'
 
-import { AuthSignUpFormType } from '@/lib/definitions'
-import { NAME_PATTERN, EMAIL_PATTERN, PASSWORD_PATTERN } from '@/lib/constants'
-import { useSignup } from '@/hooks/auth/use-signup'
-
-const SignupFormComponent = () => {
+const SigninForm = () => {
   const [passwordVisibility, setPasswordVisibility] = useState<boolean>(false)
-  const [confirmPasswordVisibility, setconfirmPasswordVisibility] = useState<boolean>(false)
 
   const {
     register,
     handleSubmit,
-    watch,
     formState: { errors }
-  } = useForm<AuthSignUpFormType>()
-  const { loading, signUp, message, success } = useSignup()
+  } = useForm<AuthSignInFormType>()
+  const { loading, signIn, message, success } = useSignin()
 
-  const password = watch('password')
-
-  const onSubmit: SubmitHandler<AuthSignUpFormType> = async (form) => await signUp(form)
+  const onSubmit: SubmitHandler<AuthSignInFormType> = async (form: AuthSignInFormType) =>
+    await signIn(form)
 
   return (
     <form className='space-y-6' onSubmit={handleSubmit(onSubmit)}>
@@ -39,25 +36,6 @@ const SignupFormComponent = () => {
         </Alert>
       ) : null}
 
-      <div>
-        <Label
-          htmlFor='name'
-          className='block text-sm font-medium text-gray-900 dark:text-gray-300'
-        >
-          Name
-        </Label>
-        <div className='mt-1'>
-          <Input
-            autoComplete='name'
-            id='name'
-            type='text'
-            placeholder='John Doe'
-            className='block w-full appearance-none rounded-md border border-gray-400 px-3 py-2 placeholder-gray-500 shadow-sm focus:border-gray-700 focus:outline-none focus:ring-gray-700 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-300 dark:placeholder-gray-500'
-            {...register('name', { required: 'Name is required', pattern: NAME_PATTERN })}
-          />
-          {errors.name && <span className='text-red-500 text-sm'>{errors.name.message}</span>}
-        </div>
-      </div>
       <div>
         <Label
           htmlFor='email'
@@ -72,7 +50,7 @@ const SignupFormComponent = () => {
             type='email'
             placeholder='john.doe@example.com'
             className='block w-full appearance-none rounded-md border border-gray-400 px-3 py-2 placeholder-gray-500 shadow-sm focus:border-gray-700 focus:outline-none focus:ring-gray-700 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-300 dark:placeholder-gray-500'
-            {...register('email', { required: 'Email is required', pattern: EMAIL_PATTERN })}
+            {...register('email', { required: 'Email is required' })}
           />
           {errors.email && <span className='text-red-500 text-sm'>{errors.email.message}</span>}
         </div>
@@ -93,8 +71,7 @@ const SignupFormComponent = () => {
               placeholder='****************'
               className='block w-full appearance-none rounded-md border border-gray-400 px-3 py-2 placeholder-gray-500 shadow-sm focus:border-gray-700 focus:outline-none focus:ring-gray-700 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-300 dark:placeholder-gray-500'
               {...register('password', {
-                required: 'Password is required',
-                pattern: PASSWORD_PATTERN
+                required: 'Password is required'
               })}
             />
             <Button
@@ -113,65 +90,46 @@ const SignupFormComponent = () => {
               <span className='sr-only'>Toggle password visibility</span>
             </Button>
           </div>
+
           {errors.password && (
             <span className='text-red-500 text-sm'>{errors.password.message}</span>
           )}
         </div>
       </div>
-      <div>
-        <Label
-          htmlFor='confirmPassword'
-          className='block text-sm font-medium text-gray-900 dark:text-gray-300'
-        >
-          Re-Type Password
-        </Label>
-        <div className='mt-1'>
-          <div className='relative'>
-            <Input
-              autoComplete='confirmPassword'
-              id='confirmPassword'
-              type={confirmPasswordVisibility ? 'text' : 'password'}
-              placeholder='****************'
-              className='block w-full appearance-none rounded-md border border-gray-400 px-3 py-2 placeholder-gray-500 shadow-sm focus:border-gray-700 focus:outline-none focus:ring-gray-700 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-300 dark:placeholder-gray-500'
-              {...register('confirmPassword', {
-                required: 'Confirm Password is required',
-                validate: (value) => value === password || 'Passwords do not match',
-                pattern: PASSWORD_PATTERN
-              })}
-            />
-            <Button
-              variant='ghost'
-              size='icon'
-              type='button'
-              className='absolute top-1/2 right-3 -translate-y-1/2 h-7 w-7 hover:bg-transparent text-gray-500 hover:text-gray-900 dark:text-gray-400 dark:hover:text-gray-300'
-              onClick={() => setconfirmPasswordVisibility(!confirmPasswordVisibility)}
-            >
-              {confirmPasswordVisibility ? (
-                <EyeOffIcon className='h-4 w-4' />
-              ) : (
-                <EyeIcon className='h-4 w-4' />
-              )}
-
-              <span className='sr-only'>Toggle password visibility</span>
-            </Button>
-          </div>
-
-          {errors.confirmPassword && (
-            <span className='text-red-500 text-sm'>{errors.confirmPassword.message}</span>
-          )}
+      <div className='flex items-center justify-between'>
+        <div className='flex items-center'>
+          <Checkbox
+            id='remember-me'
+            name='remember-me'
+            className='h-4 w-4 rounded text-gray-700 focus:ring-gray-700 dark:bg-gray-800 dark:border-gray-600 dark:checked:bg-gray-700'
+          />
+          <Label
+            htmlFor='remember-me'
+            className='ml-2 block text-sm text-gray-900 dark:text-gray-300'
+          >
+            Remember me
+          </Label>
+        </div>
+        <div className='text-sm'>
+          <Link
+            href='#'
+            className='font-medium text-gray-900 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-300'
+            prefetch={false}
+          >
+            Forgot your password?
+          </Link>
         </div>
       </div>
       <div>
         <Button
           type='submit'
-          disabled={loading}
           className='flex w-full justify-center rounded-md border border-transparent bg-gray-900 py-2 px-4 text-sm font-medium text-white shadow-sm hover:bg-gray-800 focus:outline-none focus:ring-2 focus:ring-gray-700 focus:ring-offset-2 dark:bg-gray-400 dark:hover:bg-gray-500 dark:focus:ring-gray-600'
         >
-          {loading ? 'Loading...' : 'Sign up'}
+          {loading ? 'Loading...' : 'Sign in'}
         </Button>
       </div>
     </form>
   )
 }
 
-export { SignupFormComponent }
+export { SigninForm }
