+++ components/root/footer.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import Image from 'next/image'
 
-const Footercomponent = () => {
+const Footer = () => {
   return (
     <footer className='flex flex-col gap-2 sm:flex-row py-6 w-full shrink-0 items-center px-4 md:px-6'>
       <p className='text-xs text-gray-500 dark:text-gray-400'>
@@ -17,4 +17,4 @@ const Footercomponent = () => {
   )
 }
 
-export { Footercomponent }
+export { Footer }
